requirements_frontend.txt
chromadb
streamlit
requests
azure-identity
azure-keyvault-secrets